django_backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication & Security
djangorestframework-simplejwt==5.3.0
bcrypt==4.0.1
cryptography==41.0.7

# Database & ORM
psycopg2-binary==2.9.9
mysqlclient==2.2.0
pymongo==4.6.0

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0

# Background Jobs & Caching
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# File Upload & Storage
django-storages==1.14.2
boto3==1.34.0
Pillow==10.1.0

# Communication & External APIs
requests==2.31.0
stripe==7.9.0
mailgun==0.1.1
twilio==8.11.1

# Data Processing & Analytics
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# PDF & Document Processing
reportlab==4.0.7
PyPDF2==3.0.1

# Monitoring & Logging
sentry-sdk==1.38.0
django-extensions==3.2.3

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==20.1.0
django-debug-toolbar==4.2.0

# API Documentation
drf-spectacular==0.26.5

# Mathematical Finance
scipy==1.11.4
financepy==0.34